# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dobrenok Blog API",
        "description": "API для получения проектов и постов блога. Включает пагинацию, поиск, фильтрацию и сортировку.",
        "contact": {
            "name": "API Support",
            "email": "support@dobrenok.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Local development server"
        },
        {
            "url": "https://api.dobrenok.com/api/v1",
            "description": "Production server"
        }
    ],
    "paths": {
...
        "/projects/{id}/events": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Список мероприятий проекта",
                "description": "Возвращает пагинированный список мероприятий для указанного проекта",
                "operationId": "7ce378f2ea5aaec695ae9021c9ff8b8c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID проекта",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "month",
                        "in": "query",
                        "description": "Месяц (1-12) для фильтрации по дате начала",
                        "schema": {
                            "type": "integer",
                            "maximum": 12,
                            "minimum": 1,
                            "example": 9
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Номер страницы",
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Количество элементов на странице",
                        "schema": {
                            "type": "integer",
                            "maximum": 50,
                            "minimum": 1,
                            "example": 15
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Поиск по названию или описанию",
                        "schema": {
                            "type": "string",
                            "example": "митап"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Поле для сортировки",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "id",
                                "name",
                                "start_date",
                                "created_at",
                                "updated_at"
                            ],
                            "example": "start_date"
                        }
                    },
                    {
                        "name": "sort_direction",
                        "in": "query",
                        "description": "Направление сортировки",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ],
                            "example": "desc"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Event"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "pagination": {
                                                    "$ref": "#/components/schemas/PaginationMeta"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/projects/{id}/partners": {
            "get": {
                "tags": [
                    "Partners"
                ],
                "summary": "Список партнеров проекта",
                "description": "Возвращает пагинированный список партнеров для указанного проекта",
                "operationId": "2db172aad8511ff92cf09d75cbb1160e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID проекта",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Номер страницы",
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Количество элементов на странице",
                        "schema": {
                            "type": "integer",
                            "maximum": 50,
                            "minimum": 1,
                            "example": 15
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Поиск по названию или описанию",
                        "schema": {
                            "type": "string",
                            "example": "Acme"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Поле для сортировки",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "id",
                                "name",
                                "created_at",
                                "updated_at"
                            ],
                            "example": "created_at"
                        }
                    },
                    {
                        "name": "sort_direction",
                        "in": "query",
                        "description": "Направление сортировки",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ],
                            "example": "desc"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Partner"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "pagination": {
                                                    "$ref": "#/components/schemas/PaginationMeta"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Event": {
                "description": "Мероприятие проекта",
                "properties": {
                    "id": {
                        "description": "ID мероприятия",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Название",
                        "type": "string",
                        "example": "Митап Laravel"
                    },
                    "slug": {
                        "description": "Ссылка",
                        "type": "string",
                        "example": "mitap-laravel"
                    },
                    "short_description": {
                        "description": "Краткое описание",
                        "type": "string",
                        "example": "Краткое описание мероприятия..."
                    },
                    "body": {
                        "description": "Описание мероприятия",
                        "type": "string",
                        "example": "Подробности мероприятия..."
                    },
                    "cover_photo_url": {
                        "description": "URL обложки",
                        "type": "string",
                        "example": "http://localhost:8000/storage/event-images/cover.jpg"
                    },
                    "start_date": {
                        "description": "Дата начала",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-09-20T18:00:00Z"
                    },
                    "created_at": {
                        "description": "Дата создания",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-09-19T10:30:00Z"
                    },
                    "updated_at": {
                        "description": "Дата обновления",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-09-19T10:30:00Z"
                    }
                },
                "type": "object"
            },
            "Partner": {
                "description": "Партнер проекта",
                "properties": {
                    "id": {
                        "description": "ID партнера",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Название партнера",
                        "type": "string",
                        "example": "Acme Inc."
                    },
                    "description": {
                        "description": "Описание партнера",
                        "type": "string",
                        "example": "Описание партнера",
                        "nullable": true
                    },
                    "cover_photo_url": {
                        "description": "URL изображения",
                        "type": "string",
                        "example": "http://localhost:8000/storage/partner-images/logo.png"
                    },
                    "created_at": {
                        "description": "Дата создания",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15T10:30:00Z"
                    },
                    "updated_at": {
                        "description": "Дата обновления",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15T10:30:00Z"
                    }
                },
                "type": "object"
            },
...
            "PaginationMeta": {
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 5
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 15
                    },
                    "total": {
                        "type": "integer",
                        "example": 75
                    },
                    "next_page_url": {
                        "type": "string",
                        "example": "http://localhost:8000/api/v1/projects?page=2",
                        "nullable": true
                    },
                    "prev_page_url": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "ErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "error": {
                        "properties": {
                            "type": {
                                "type": "string",
                                "example": "not_found"
                            },
                            "message": {
                                "type": "string",
                                "example": "Resource not found"
                            },
                            "details": {
                                "type": "object",
                                "nullable": true
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}